# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/journal/audit": {
      "get": {
        "description": "Toute écriture y figure. Les libellés d’agent et de cible sont recalculés à la lecture, jamais recopiés à l’écriture — une trace figée survivrait à une anonymisation.",
        "operationId": "JournalController_audit",
        "parameters": [
          {
            "name": "agent",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cible",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "decalage",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "limite",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntreeAuditDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Journal d’audit",
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/consultations": {
      "get": {
        "description": "Toute lecture de fiche, **y compris celle d’un super-admin**, qui y est signalée comme telle. `derogation` distingue la lecture permise par habilitation de celle qui n’a été possible que par dérogation.",
        "operationId": "JournalController_consultations",
        "parameters": [
          {
            "name": "agent",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "objet",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "superAdmin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "derogation",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decalage",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "limite",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntreeConsultationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Journal de consultation",
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/orphelines": {
      "get": {
        "description": "Entités sans aucun lien actif, qu’une saisie en cascade interrompue peut laisser derrière elle. Liste discrète : elles ne remontent jamais en signal sur l’accueil — ce n’est pas un rapprochement, c’est du ménage.",
        "operationId": "JournalController_orphelines",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntiteOrphelineDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Entités orphelines",
        "tags": [
          "journal"
        ]
      }
    },
...
    "/accueil": {
      "get": {
        "description": "Signaux, dossiers de l’agent et dernière activité en une requête. Tout y est calculé **après** filtrage de visibilité : un signal portant sur un objet inaccessible ne remonte pas, sa seule mention en révélerait l’existence.",
        "operationId": "SignauxController_accueil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccueilDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Écran d’accueil, assemblé",
        "tags": [
          "accueil"
        ]
      }
    },
    "/signaux": {
      "get": {
        "description": "Recoupement, récurrence et vieillissement. Les trois familles se calculent sur ce que cet agent voit, et sur rien d’autre.",
        "operationId": "SignauxController_liste",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Signaux seuls",
        "tags": [
          "accueil"
        ]
      }
    },
    "/recherche": {
      "get": {
        "description": "Entités et dossiers. Les objets inaccessibles en sont absents, sans mention ni décompte : un total qui ne tombe pas juste est déjà une information.",
        "operationId": "SignauxController_rechercher",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Deux caractères au moins",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultatRechercheDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Recherche globale",
        "tags": [
          "accueil"
        ]
      }
    },
...
  },
  "info": {
    "title": "Centrale N&I",
    "description": "API du service Narcotics & Investigations de la SAHP. Le référentiel — types d'entités, champs, types de liens, onglets — est dynamique et n'apparaît pas dans ce contrat : il se récupère par GET /referentiel.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jeton": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "EntreeAuditDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant de l’entrée, sérialisé en texte"
          },
          "agentId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "agentLibelle": {
            "type": "string",
            "description": "« agent supprimé » si le compte est anonymisé"
          },
          "action": {
            "type": "string"
          },
          "cibleTable": {
            "type": "string"
          },
          "cibleId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "cibleLibelle": {
            "type": "string",
            "nullable": true,
            "description": "Libellé de la cible, lorsqu’elle est retrouvable"
          },
          "avant": {
            "type": "object",
            "additionalProperties": true
          },
          "apres": {
            "type": "object",
            "additionalProperties": true
          },
          "effectueLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "agentId",
          "agentLibelle",
          "action",
          "cibleTable",
          "cibleId",
          "cibleLibelle",
          "effectueLe"
        ]
      },
      "EntreeConsultationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "agentLibelle": {
            "type": "string"
          },
          "nature": {
            "enum": [
              "entite",
              "dossier"
            ],
            "type": "string"
          },
          "objetId": {
            "type": "string",
            "format": "uuid"
          },
          "objetLibelle": {
            "type": "string",
            "nullable": true
          },
          "derogation": {
            "type": "boolean",
            "description": "Lecture rendue possible par une permission dérogatoire"
          },
          "superAdmin": {
            "type": "boolean",
            "description": "Lecture d’un super-admin, signalée comme telle"
          },
          "consulteLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "agentId",
          "agentLibelle",
          "nature",
          "objetId",
          "objetLibelle",
          "derogation",
          "superAdmin",
          "consulteLe"
        ]
      },
      "EntiteOrphelineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "creeLe": {
            "type": "string",
            "format": "date-time"
          },
          "auteur": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "libelle",
          "typeCode",
          "creeLe",
          "auteur"
        ]
      },
...
      "SignalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Clé stable, pour que le front puisse suivre un signal"
          },
          "famille": {
            "enum": [
              "recoupement",
              "recurrence",
              "vieillissement"
            ],
            "type": "string"
          },
          "entiteId": {
            "type": "string",
            "format": "uuid"
          },
          "entiteLibelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "resume": {
            "type": "string",
            "description": "Ce que la centrale a remarqué, en une phrase"
          },
          "detail": {
            "type": "string",
            "description": "Ce sur quoi elle s’appuie"
          },
          "faitId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Le fait en cause, pour le vieillissement"
          }
        },
        "required": [
          "id",
          "famille",
          "entiteId",
          "entiteLibelle",
          "typeCode",
          "resume",
          "detail",
          "faitId"
        ]
      },
      "DossierDeLAgentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nom": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "entitePivotId": {
            "type": "string",
            "format": "uuid"
          },
          "entitePivotLibelle": {
            "type": "string"
          },
          "motif": {
            "enum": [
              "habilitation",
              "creation"
            ],
            "type": "string",
            "description": "Habilité nommément, ou auteur du dossier"
          }
        },
        "required": [
          "id",
          "nom",
          "visibilite",
          "entitePivotId",
          "entitePivotLibelle",
          "motif"
        ]
      },
      "ActiviteDto": {
        "type": "object",
        "properties": {
          "faitId": {
            "type": "string",
            "format": "uuid"
          },
          "entiteId": {
            "type": "string",
            "format": "uuid"
          },
          "entiteLibelle": {
            "type": "string"
          },
          "resume": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "auteur": {
            "type": "string",
            "nullable": true,
            "description": "« agent supprimé » si anonymisé"
          },
          "survenuLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "faitId",
          "entiteId",
          "entiteLibelle",
          "resume",
          "source",
          "fiabilite",
          "auteur",
          "survenuLe"
        ]
      },
      "AccueilDto": {
        "type": "object",
        "properties": {
          "signaux": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignalDto"
            }
          },
          "mesDossiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DossierDeLAgentDto"
            }
          },
          "derniereActivite": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActiviteDto"
            }
          }
        },
        "required": [
          "signaux",
          "mesDossiers",
          "derniereActivite"
        ]
      },
      "ResultatRechercheDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "nature": {
            "enum": [
              "entite",
              "dossier"
            ],
            "type": "string"
          },
          "typeCode": {
            "type": "string",
            "nullable": true
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "libelle",
          "nature",
          "typeCode",
          "visibilite"
        ]
      },
...
    }
  }
}